Updates mattermost/server/public & removes telemetry (#952)
