+++ server/plugin/flows.go
@@ -25,11 +25,6 @@ type PingBroker interface {
 	SubscribePings() <-chan *github.PingEvent
 }
 
-type Tracker interface {
-	TrackEvent(event string, properties map[string]interface{})
-	TrackUserEvent(event, userID string, properties map[string]interface{})
-}
-
 type FlowManager struct {
 	client           *pluginapi.Client
 	pluginID         string
@@ -40,7 +35,6 @@ type FlowManager struct {
 	useGitHubClient  func(info *GitHubUserInfo, toRun func(info *GitHubUserInfo, token *oauth2.Token) error) error
 
 	pingBroker PingBroker
-	tracker    Tracker
 
 	setupFlow        *flow.Flow
 	oauthFlow        *flow.Flow
@@ -59,7 +53,6 @@ func (p *Plugin) NewFlowManager() (*FlowManager, error) {
 		useGitHubClient:  p.useGitHubClient,
 
 		pingBroker: p.webhookBroker,
-		tracker:    p,
 	}
 
 	setupFlow, err := fm.newFlow("setup")
@@ -141,8 +134,6 @@ func (fm *FlowManager) doneStep() flow.Step {
 }
 
 func (fm *FlowManager) onDone(f *flow.Flow) {
-	fm.trackCompleteSetupWizard(f.UserID)
-
 	delegatedFrom := f.GetState().GetString(keyDelegatedFrom)
 	if delegatedFrom != "" {
 		err := fm.setupFlow.ForUser(delegatedFrom).Go(stepDelegateComplete)
@@ -252,22 +243,7 @@ func (fm *FlowManager) StartSetupWizard(userID string, delegatedFrom string) err
 
 	fm.client.Log.Debug("Started setup wizard", "userID", userID, "delegatedFrom", delegatedFrom)
 
-	fm.trackStartSetupWizard(userID, delegatedFrom != "")
-
 	return nil
-}
-
-func (fm *FlowManager) trackStartSetupWizard(userID string, fromInvite bool) {
-	fm.tracker.TrackUserEvent("setup_wizard_start", userID, map[string]interface{}{
-		"from_invite": fromInvite,
-		"time":        model.GetMillis(),
-	})
-}
-
-func (fm *FlowManager) trackCompleteSetupWizard(userID string) {
-	fm.tracker.TrackUserEvent("setup_wizard_complete", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
 }
 
 func (fm *FlowManager) StartOauthWizard(userID string) error {
@@ -278,21 +254,7 @@ func (fm *FlowManager) StartOauthWizard(userID string) error {
 		return err
 	}
 
-	fm.trackStartOauthWizard(userID)
-
 	return nil
-}
-
-func (fm *FlowManager) trackStartOauthWizard(userID string) {
-	fm.tracker.TrackUserEvent("oauth_wizard_start", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
-}
-
-func (fm *FlowManager) trackCompleteOauthWizard(userID string) {
-	fm.tracker.TrackUserEvent("oauth_wizard_complete", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
 }
 
 func (fm *FlowManager) stepWelcome() flow.Step {
@@ -616,8 +578,7 @@ func (fm *FlowManager) stepOAuthConnect() flow.Step {
 	connectText := fmt.Sprintf("Go [here](%s) to connect your account.", connectURL)
 	return flow.NewStep(stepOAuthConnect).
 		WithText(connectText).
-		WithPretext(connectPretext).
-		OnRender(func(f *flow.Flow) { fm.trackCompleteOauthWizard(f.UserID) })
+		WithPretext(connectPretext)
 	// The API handler will advance to the next step and complete the flow
 }
 
@@ -629,21 +590,7 @@ func (fm *FlowManager) StartWebhookWizard(userID string) error {
 		return err
 	}
 
-	fm.trackStartWebhookWizard(userID)
-
 	return nil
-}
-
-func (fm *FlowManager) trackStartWebhookWizard(userID string) {
-	fm.tracker.TrackUserEvent("webhook_wizard_start", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
-}
-
-func (fm *FlowManager) trackCompleteWebhookWizard(userID string) {
-	fm.tracker.TrackUserEvent("webhook_wizard_complete", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
 }
 
 func (fm *FlowManager) stepWebhookQuestion() flow.Step {
@@ -799,7 +746,6 @@ func (fm *FlowManager) stepWebhookConfirmation() flow.Step {
 	return flow.NewStep(stepWebhookConfirmation).
 		WithTitle("Success! :tada: You've successfully set up your Mattermost GitHub integration! ").
 		WithText("Use `/github subscriptions add` to subscribe any Mattermost channel to your GitHub repository. [Learn more](https://github.com/mattermost/mattermost-plugin-github#slash-commands)").
-		OnRender(func(f *flow.Flow) { fm.trackCompleteWebhookWizard(f.UserID) }).
 		Next("")
 }
 
@@ -811,21 +757,7 @@ func (fm *FlowManager) StartAnnouncementWizard(userID string) error {
 		return err
 	}
 
-	fm.trackStartAnnouncementWizard(userID)
-
 	return nil
-}
-
-func (fm *FlowManager) trackStartAnnouncementWizard(userID string) {
-	fm.tracker.TrackUserEvent("announcement_wizard_start", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
-}
-
-func (fm *FlowManager) trackCompletAnnouncementWizard(userID string) {
-	fm.tracker.TrackUserEvent("announcement_wizard_complete", userID, map[string]interface{}{
-		"time": model.GetMillis(),
-	})
 }
 
 func (fm *FlowManager) stepAnnouncementQuestion() flow.Step {
@@ -909,8 +841,7 @@ func (fm *FlowManager) submitChannelAnnouncement(f *flow.Flow, submitted map[str
 func (fm *FlowManager) stepAnnouncementConfirmation() flow.Step {
 	return flow.NewStep(stepAnnouncementConfirmation).
 		WithText("Message to ~{{ .ChannelName }} was sent.").
-		Next("").
-		OnRender(func(f *flow.Flow) { fm.trackCompletAnnouncementWizard(f.UserID) })
+		Next("")
 }
 
 func printGithubErrorResponse(err *github.ErrorResponse) error {
