+++ server/plugin/support_packet.go
@@ -5,6 +5,7 @@ package plugin
 
 import (
 	"path/filepath"
+	"strings"
 
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
@@ -12,6 +13,11 @@ import (
 
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/plugin"
+	"github.com/mattermost/mattermost/server/public/pluginapi"
+)
+
+const (
+	keysPerPage = 1000
 )
 
 type SupportPacket struct {
@@ -45,4 +51,27 @@ func (p *Plugin) GenerateSupportData(_ *plugin.Context) ([]*model.FileData, erro
 		Filename: filepath.Join(Manifest.Id, "diagnostics.yaml"),
 		Body:     body,
 	}}, result.ErrorOrNil()
+}
+
+func (p *Plugin) getConnectedUserCount() (int64, error) {
+	checker := func(key string) (keep bool, err error) {
+		return strings.HasSuffix(key, githubTokenKey), nil
+	}
+
+	var count int64
+
+	for i := 0; ; i++ {
+		keys, err := p.store.ListKeys(i, keysPerPage, pluginapi.WithChecker(checker))
+		if err != nil {
+			return 0, errors.Wrapf(err, "failed to list keys - page, %d", i)
+		}
+
+		count += int64(len(keys))
+
+		if len(keys) < keysPerPage {
+			break
+		}
+	}
+
+	return count, nil
 }
